[9.0][MIG] connector_prestashop_catalog_manager: Migration and new structure
* This is a combination of 6 commits.

[9.0][MIG] connector_prestashop_catalog_manager: Migration and new structure
[9.0][FIX] connector_prestashop_catalog_manager: Fix export impact price
[9.0][FIX] connector_prestashop_catalog_manager: Fix export default combination
[9.0][FIX] custom_connector_prestashop_catalog_feature: Remove default_combination from product template export mapper
[9.0][FIX] connector_prestashop_catalog_manager: Fix tax group in product_template export mapper
[9.0][FIX] connector_prestashop_catalog_manager: Fix error when export a product without taxes
