+++ connector_prestashop_catalog_manager/__openerp__.py
@@ -21,10 +21,10 @@
     "data": [
         'views/product_attribute_view.xml',
         'views/product_view.xml',
-        'wizard/export_category_view.xml',
-        'wizard/export_multiple_products_view.xml',
-        'wizard/sync_products_view.xml',
-        'wizard/active_deactive_products_view.xml',
+        'wizards/export_category_view.xml',
+        'wizards/export_multiple_products_view.xml',
+        'wizards/sync_products_view.xml',
+        'wizards/active_deactive_products_view.xml',
         'views/product_image_view.xml',
     ],
     "installable": True,
