+++ connector_prestashop_catalog_manager/models/__init__.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+# © 2016 Sergio Teruel <sergio.teruel@tecnativa.com>
+# License AGPL-3 - See http://www.gnu.org/licenses/agpl-3.0.html
+
+from . import product_category
+from . import product_product
+from . import product_template
+from . import product_image
