+++ connector_prestashop_catalog_manager/models/product_product/common.py
@@ -1,0 +1,14 @@
+# -*- coding: utf-8 -*-
+# © 2016 Sergio Teruel <sergio.teruel@tecnativa.com>
+# License AGPL-3 - See http://www.gnu.org/licenses/agpl-3.0.html
+
+from openerp import models, fields
+
+
+class PrestashopProductCombination(models.Model):
+    _inherit = 'prestashop.product.combination'
+    minimal_quantity = fields.Integer(
+        string='Minimal Quantity',
+        default=1,
+        help='Minimal Sale quantity',
+    )
