+++ connector_prestashop_catalog_manager/wizards/__init__.py
