+++ connector_prestashop_catalog_manager/wizards/active_deactive_products_view.xml
