+++ connector_prestashop_catalog_manager/wizards/export_category.py
