+++ connector_prestashop_catalog_manager/wizards/sync_products_view.xml
